fix: Fix blogs page
