+++ src/components/atoms/FloatingButton/type.ts
@@ -1,4 +1,6 @@
-export interface Props {
+import { ButtonProps } from '@chakra-ui/react'
+
+export interface Props extends ButtonProps {
   onClick: (args?: unknown) => void
   processing: boolean
 }
