+++ src/components/molecules/BlogSearchForm/type.ts
@@ -1,6 +1,6 @@
 export interface Props {
   clear: () => void
   value: string
-  onChange: () => void
-  onKeyPress: () => void
+  onChange: (event?: any) => void
+  onKeyPress: (event?: any) => void
 }
