+++ src/components/templates/blog/BlogsTemplate.tsx
@@ -4,16 +4,14 @@ import {
   Flex,
   useBreakpointValue,
   useColorModeValue,
-  Center,
 } from '@chakra-ui/react'
 import { InlineResponse400, InlineResponse2002 } from '@/apis/models'
-// import Ads from '@/components/organisms/Ads'
-import { LoadingIconComponent } from '@/components/atoms/LoadingIcon/LoadingIcon'
-import BlogsContents from '@/components/organisms/BlogsContents'
-import BlogsSideMenu from '@/components/organisms/BlogsSideMenu'
+import BlogsContents from '@/components/features/blogs/BlogsContents'
+import BlogsSideMenu from '@/components/features/blogs/BlogsSideMenu'
+import LoadingBlogContents from '@/components/features/blogs/LoadingBlogContents'
+import SideMenuWrapper from '@/components/features/blogs/SideMenuWrapper'
 import { FULL_HEIGHT } from '@/consts/style'
 import useDesignSystem from '@/hooks/useDesignSystem'
-import useSp from '@/hooks/useSp'
 import useWindowHeight from '@/hooks/useWindowHeight'
 import Error from '@/pages/_error'
 import { PageWrapper } from '@/styles/global.css'
@@ -29,16 +27,11 @@ interface Props {
 
 const BlogsTemplate: React.FC<Props> = memo(
   ({ settings, contents, title, searchWord, isLoading, isError }) => {
-    const [isSp] = useSp()
     const { BLOG_SIDE_MENU_BG } = useDesignSystem()
     const { scrollHeight } = useWindowHeight()
     const minHeight = useBreakpointValue({
       base: 'auto',
       md: scrollHeight,
-    })
-    const sideMenuDisplay = useBreakpointValue({
-      base: 'none',
-      md: 'inherit',
     })
     const flexFlow = useBreakpointValue({
       base: 'column-reverse',
@@ -67,15 +60,13 @@ const BlogsTemplate: React.FC<Props> = memo(
     return (
       <PageWrapper id="blog-container">
         <Flex flexFlow={flexFlow} justifyContent="space-between">
-          {/* メニュー */}
-          {/* スマホサイズじゃない場合 */}
-          {!isSp && (
+          {/* サイドメニュー */}
+          <SideMenuWrapper>
             <Box
               maxW={sideMenuMaxW}
               w={sideMenuW}
               minH={minHeight}
               position="relative"
-              bgColor={BLOG_SIDE_MENU_BG}
             >
               <Box
                 minH={FULL_HEIGHT}
@@ -84,20 +75,17 @@ const BlogsTemplate: React.FC<Props> = memo(
                 left={0}
                 w="100%"
               >
-                <BlogsSideMenu data={(settings as InlineResponse2002).data} />
+                <BlogsSideMenu
+                  data={(settings as InlineResponse2002).data}
+                  loading={isLoading}
+                />
               </Box>
             </Box>
-          )}
+          </SideMenuWrapper>
           {/* コンテンツ */}
-          <Box w={contentW} minH={minHeight}>
+          <Box w={contentW} minH={minHeight} flex={1}>
             {/* 取得中 */}
-            {isLoading && (
-              <Center w="full" h="full">
-                <div className={`loader ${colorTheme}`}>
-                  <LoadingIconComponent />
-                </div>
-              </Center>
-            )}
+            {isLoading && <LoadingBlogContents />}
             {/* コンテンツが返ってきた場合 */}
             {!isLoading && (
               <BlogsContents
@@ -107,16 +95,6 @@ const BlogsTemplate: React.FC<Props> = memo(
                 searchWord={searchWord}
               />
             )}
-          </Box>
-          {/* サイドメニュー */}
-          <Box
-            as="aside"
-            flex={1}
-            minH={minHeight}
-            display={sideMenuDisplay}
-            maxW="360px"
-          >
-            {/* <Ads /> */}
           </Box>
         </Flex>
       </PageWrapper>
