+++ src/hooks/useDesignSystem.ts
@@ -10,7 +10,7 @@ type UseDesignSystem = () => {
 
 const useDesignSystem: UseDesignSystem = () => {
   const isDark = useColorModeValue(false, true)
-  const NO_DATA_COLOR = useColorModeValue('app-gray.200', 'app-gray.800')
+  const NO_DATA_COLOR = useColorModeValue('app-gray.800', 'app-gray.200')
   const BLOG_SIDE_MENU_BG = useColorModeValue('app-gray.200', 'app-gray.800')
   const TEXT_COLOR = useColorModeValue('app-gray.900', 'app-gray.50')
   const BG_COLOR = useColorModeValue('app-gray.50', 'app-gray.700')
